requirements.txt
google-adk==0.3.0
google-generativeai>=0.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
